[Fix] Address Step 4 review: ping-pong buffers, obs dtype, import cleanup
- Fix ping-pong race on _cpu_buf: use two alternating CPU buffers in
  ObservationBuffer.copy_to_cpu() so overlap scheduling never corrupts
  the previous step's observation data
- Fix obs_mask dtype mismatch: cast per_token_mask to x.dtype in
  observe() to prevent index_copy_ crash with bfloat16 models
- Move get_intervention_ctx import to module level in engine.py (hot path)
- Fix assert style: is False → == False for consistency (args.py)
- Update tests for ping-pong behavior and add bfloat16 regression test

+++ python/minisgl/intervention/ops.py
@@ -25,7 +25,7 @@ def observe(
         base_indices: [max_tokens_per_slot] pre-allocated [0, 1, 2, ...]
         offsets: [num_layers] pre-computed layer offsets into flat_buf
     """
-    per_token_mask = obs_mask[layer_idx, req_map]  # [n_tokens]
+    per_token_mask = obs_mask[layer_idx, req_map].to(dtype=x.dtype)  # [n_tokens]
     masked = x * per_token_mask.unsqueeze(-1)  # [n_tokens, hidden_dim]
     indices = base_indices[: x.shape[0]] + offsets[layer_idx]
     flat_buf.index_copy_(0, indices, masked)
